requirements.txt.txt
streamlit
azure-storage-blob
pandas
matplotlib
python-dotenv
reportlab
requests